backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
supabase==2.3.0
openai==1.6.0
web3==6.11.3
pydantic==2.5.3
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4
